Allow specifying a custom transition name for JIRA auto start
If you want to set your issue to something other than 'Start Progress' you
currently have to edit the source code. Now it is possible to use the
auto_start config value for this purpose.

The old y/n values are still fully supported and will continue to work as
before.